requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.13
django-ratelimit==4.1.0
gunicorn==25.3.0
packaging==26.2
psycopg==3.3.3
psycopg-binary==3.3.3
python-decouple==3.8
ruff==0.15.12
sqlparse==0.5.5
whitenoise==6.12.0
